Fix: API 분리
